+++ changelog.md
@@ -7,6 +7,7 @@ vX.X.X
 - TX and RX buffer sizes are now shown on stats sub-tab, closes #95.
 - Fixed bug where build configurations were not being included in repo, closes #85.
 - Added logging sub-tab and basic logging functionality, closes #96.
+- Greyed out textfield and browse button on logging tab when logging is active, closes #97.
 
 v0.4.1
 ------